# Improve performance of Sort for the common single batch use case