--- a/sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuSortExec.scala
+++ b/sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuSortExec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2023, NVIDIA CORPORATION.
+ * Copyright (c) 2019-2024, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -463,7 +463,7 @@ case class GpuOutOfCoreSortIterator(
    */
   private final def firstPassReadBatches(scb: SpillableColumnarBatch): Unit = {
     splitOneSortedBatch(scb)
-    while(alreadySortedIter.hasNext) {
+    while (alreadySortedIter.hasNext) {
       splitOneSortedBatch(alreadySortedIter.next())
     }
   }